Create WebScrapingCustomised
Customised web scraping for iteration over local HTML files organised into folders for separate DD, YY and MM.
This script won't run as-is, given that HTML tags and attributes used varies from site to site and XMLtree to XMLtree.  xPath is used to extract and remove specific classes + attributes, thus the path in the XMLtree of the page to be scraped will need to be examined to determine precisely what should be entered as the xpath, these are identified with <<text>> in xpath expressions  above.

It may not be necessary to extract/remove specific elements from some pages, whereas other elements not identified here may exist on other pages.